Fix and Update JavaScript Recourses Links (#5896)
* Update JavaScript-asynchronous Promises resources

* Replace broken link

The previous link to the article was outdated and resulted in a 404 error.

* Fixed the JavaScript typo

* Add Video Resource JS 101-debugging-memory-leaks.md

Added a helpful YouTube video on visualizing memory leaks and debugging them in the console.

* Introduce resource (Chrome Developers) JS-debugging-performance

Added link to Chrome Developers documentation on debugging JavaScript performance in JS 102-debugging-performance.md. This provides valuable insights and best practices for optimizing code performance from the official docs.